refactor(infra): extract SaaS relay to dedicated handler (ADR-096)
Why: socket.service.ts atteignait 1263 lignes (>3× la limite 400), mélangeant
le protocole Pi ↔ Cloud avec la logique SaaS. Audit Lead Dev 2026-04-25 a
identifié ce fichier comme P1. 9 handlers Pi avaient déjà été extraits en
Phase 7.2 — la logique SaaS restait inlinée.

What:
- Nouveau handler `central-server/src/handlers/saas-relay.handler.ts` (10ème
  handler) qui contient toute la logique SaaS : registerSaasRelay,
  getSaasConnectedDisplays, sweepOrphanSaasStates, auditRemoteCommand
  (helper privé), saasStates Map + saasRelayRegistered Set en module state.
- socket.service.ts : 1263 → 991 lignes (-272 / -22%). Les 3 méthodes
  publiques/privées préservées comme wrappers qui délèguent au handler —
  surface API inchangée pour les 14 consommateurs.
- Smoke tests mis à jour (smoke-adr-refactoring, smoke-socket-realtime,
  smoke-scoreboard-saas) pour lire le handler file au lieu de socket.service.
- Test unitaire socket.service.test.ts : utilise __getSaasStatesMap() /
  __getSaasRelayRegistered() exposés par le handler comme test helpers.
- Invariants P0 préservés : sweep périodique GC + métrique Prometheus
  (PR #600) déplacés dans le handler avec le reste de la logique SaaS.
- ADR-096 léger documenté + README ADR mis à jour.

Tests: 3404/3404 verts (full Jest suite), TypeScript strict OK.
