+++ central-server/src/__tests__/smoke/smoke-socket-realtime.test.ts
@@ -430,12 +430,15 @@ describe('Deployment repository query safety', () => {
   });
 });
 
-describe('SaaS state memory leak regression guards (issue #594)', () => {
+describe('SaaS state memory leak regression guards (issue #594 + ADR-096)', () => {
   // Guards added 2026-04-25 — saasStates Map was never cleaned up on disconnect,
-  // causing unbounded growth for long-lived processes. Fixed in this session.
+  // causing unbounded growth for long-lived processes. Fixed in PR #600.
+  // ADR-096 (2026-04-25 — same day) : la logique SaaS a été extraite dans
+  // handlers/saas-relay.handler.ts. Les assertions ci-dessous lisent désormais
+  // ce handler-là (au lieu de socket.service.ts qui ne contient plus la logique).
   const repoRoot = path.resolve(__dirname, '..', '..', '..', '..');
-  const socketContent = fs.readFileSync(
-    path.join(repoRoot, 'central-server/src/services/socket.service.ts'),
+  const handlerContent = fs.readFileSync(
+    path.join(repoRoot, 'central-server/src/handlers/saas-relay.handler.ts'),
     'utf8'
   );
   const serverContent = fs.readFileSync(
@@ -446,7 +449,7 @@ describe('SaaS state memory leak regression guards (issue #594)', () => {
   it('saasStates must be deleted when last SaaS client disconnects', () => {
     // Without this, each unique siteId adds a permanent entry that is never freed.
     expect({
-      hasDeleteCall: socketContent.includes('saasStates.delete(siteId)'),
+      hasDeleteCall: handlerContent.includes('saasStates.delete(siteId)'),
     }).toEqual({ hasDeleteCall: true });
   });
 
@@ -454,20 +457,20 @@ describe('SaaS state memory leak regression guards (issue #594)', () => {
     // The delete must be conditional on tvInstances being empty AND the Socket.IO room
     // being empty, to avoid premature cleanup when other clients are still connected.
     expect({
-      guardedByTvInstances: socketContent.includes('tvInstances.size === 0'),
-      guardedByRoomSize: socketContent.includes('room.size === 0'),
-      hasDeleteCall: socketContent.includes('saasStates.delete(siteId)'),
+      guardedByTvInstances: handlerContent.includes('tvInstances.size === 0'),
+      guardedByRoomSize: handlerContent.includes('room.size === 0'),
+      hasDeleteCall: handlerContent.includes('saasStates.delete(siteId)'),
     }).toEqual({ guardedByTvInstances: true, guardedByRoomSize: true, hasDeleteCall: true });
   });
 
   it('saasStates size must be reported to Prometheus on set and delete', () => {
     // Allows Grafana to detect a growing saasStates Map before it causes OOM.
     // We verify recordSaasStatesCount is called at least twice: once on set, once on delete.
-    const occurrences = (socketContent.match(/metricsService\.recordSaasStatesCount/g) || []).length;
-    const setIdx = socketContent.indexOf('saasStates.set(siteId');
-    const deleteIdx = socketContent.indexOf('saasStates.delete(siteId)');
-    const metricAfterSet = socketContent.indexOf('recordSaasStatesCount', setIdx) < deleteIdx;
-    const metricAfterDelete = socketContent.indexOf('recordSaasStatesCount', deleteIdx) > deleteIdx;
+    const occurrences = (handlerContent.match(/metricsService\.recordSaasStatesCount/g) || []).length;
+    const setIdx = handlerContent.indexOf('saasStates.set(siteId');
+    const deleteIdx = handlerContent.indexOf('saasStates.delete(siteId)');
+    const metricAfterSet = handlerContent.indexOf('recordSaasStatesCount', setIdx) < deleteIdx;
+    const metricAfterDelete = handlerContent.indexOf('recordSaasStatesCount', deleteIdx) > deleteIdx;
     expect({
       calledAtLeastTwice: occurrences >= 2,
       metricAfterSet,
